Made changes to table and made it responsive
